% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{full_plot}
\alias{full_plot}
\title{Title}
\usage{
full_plot(save = T)
}
\arguments{
\item{save}{Boolean for whether this should save a PDF of the figure.}
}
\value{
Object with all the plots assembled.
}
\description{
Title
}
