% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{getSortingPosterior}
\alias{getSortingPosterior}
\title{Produce a plot showing the posteriors for each sorting function across the in vivo models.}
\usage{
getSortingPosterior()
}
\value{
Plot object
}
\description{
Produce a plot showing the posteriors for each sorting function across the in vivo models.
}
