% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halfl.R
\name{halfl_fcrn}
\alias{halfl_fcrn}
\title{Solves for the model half-life given a list of paraemeters.}
\usage{
halfl_fcrn(th)
}
\arguments{
\item{th}{List of parameters with which to solve.}
}
\value{
The half-life given this parameter set.
}
\description{
Solves for the model half-life given a list of paraemeters.
}
