% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halfl.R
\name{jacFunc}
\alias{jacFunc}
\title{Builds the jacobian of the ODE model.}
\usage{
jacFunc(p)
}
\arguments{
\item{p}{List of parameters with which to solve.}
}
\value{
Jacobian of the ODE system.
}
\description{
Builds the jacobian of the ODE model.
}
