% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halfl.R
\name{loadsample}
\alias{loadsample}
\title{Title}
\usage{
loadsample(name, as.df = F)
}
\arguments{
\item{name}{Name of the in vivo model fit to load.}

\item{as.df}{Do you want the return to be a data.frame?}
}
\value{
Loaded samples from a fit
}
\description{
Title
}
