% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_halfls}
\alias{plot_halfls}
\title{importFrom magrittr "%>%"
Title}
\usage{
plot_halfls(name)
}
\arguments{
\item{fit}{The rstan fit object with all the samples.}
}
\value{
The ggplot2 object.
}
\description{
importFrom magrittr "%>%"
Title
}
